+++ cpp/lifting.h
@@ -452,16 +452,24 @@ void apply_lifting(
     throw std::runtime_error(
         "Mismatch in size between a and bcs in assembler.");
   }
+  // If all forms are null, there is nothing to do
+  if (std::ranges::all_of(a, [](auto ai) { return !ai; }))
+    return;
+
   for (std::size_t j = 0; j < a.size(); ++j)
   {
-    if (x0.empty())
+    if (a[j] and !bcs1[j].empty())
     {
-      impl::apply_lifting<double>(b, a[j], bcs1[j], std::span<const double>(),
-                                  scale, mpc);
-    }
-    else
-    {
-      impl::apply_lifting<double>(b, a[j], bcs1[j], x0[j], scale, mpc);
+
+      if (x0.empty())
+      {
+        impl::apply_lifting<double>(b, a[j], bcs1[j], std::span<const double>(),
+                                    scale, mpc);
+      }
+      else
+      {
+        impl::apply_lifting<double>(b, a[j], bcs1[j], x0[j], scale, mpc);
+      }
     }
   }
 }
@@ -510,20 +518,24 @@ void apply_lifting(
         "Mismatch in size between a and bcs in assembler.");
   }
   for (std::size_t j = 0; j < a.size(); ++j)
-  {
+  { 
+    if (a[j] and !bcs1[j].empty())
+    {
     if (x0.empty())
-    {
-      impl::apply_lifting<std::complex<double>>(
-          b, a[j], bcs1[j], std::span<const std::complex<double>>(), scale,
-          mpc);
-    }
-    else
-    {
-      impl::apply_lifting<std::complex<double>>(b, a[j], bcs1[j], x0[j], scale,
-                                                mpc);
+      {
+        impl::apply_lifting<std::complex<double>>(
+            b, a[j], bcs1[j], std::span<const std::complex<double>>(), scale,
+            mpc);
+      }
+      else
+      {
+        impl::apply_lifting<std::complex<double>>(b, a[j], bcs1[j], x0[j], scale,
+                                                  mpc);
+      }
     }
   }
 }
+
 
 /// Modify b such that:
 ///
@@ -567,14 +579,17 @@ void apply_lifting(
   }
   for (std::size_t j = 0; j < a.size(); ++j)
   {
-    if (x0.empty())
+    if (a[j] and !bcs1[j].empty())
     {
-      impl::apply_lifting<float>(b, a[j], bcs1[j], std::span<const float>(),
-                                 scale, mpc);
-    }
-    else
-    {
-      impl::apply_lifting<float>(b, a[j], bcs1[j], x0[j], scale, mpc);
+      if (x0.empty())
+      {
+        impl::apply_lifting<float>(b, a[j], bcs1[j], std::span<const float>(),
+                                  scale, mpc);
+      }
+      else
+      {
+        impl::apply_lifting<float>(b, a[j], bcs1[j], x0[j], scale, mpc);
+      }
     }
   }
 }
@@ -625,15 +640,18 @@ void apply_lifting(
   }
   for (std::size_t j = 0; j < a.size(); ++j)
   {
-    if (x0.empty())
+    if (a[j] and !bcs1[j].empty())
     {
-      impl::apply_lifting<std::complex<float>>(
-          b, a[j], bcs1[j], std::span<const std::complex<float>>(), scale, mpc);
-    }
-    else
-    {
-      impl::apply_lifting<std::complex<float>>(b, a[j], bcs1[j], x0[j], scale,
-                                               mpc);
+      if (x0.empty())
+      {
+        impl::apply_lifting<std::complex<float>>(
+            b, a[j], bcs1[j], std::span<const std::complex<float>>(), scale, mpc);
+      }
+      else
+      {
+        impl::apply_lifting<std::complex<float>>(b, a[j], bcs1[j], x0[j], scale,
+                                                mpc);
+      }
     }
   }
 }