+++ python/tests/test_nonlinear_assembly.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2022 Nathan Sime
+# Copyright (C) 2022-2025 Nathan Sime, Jørgen S. Dokken
 #
 # This file is part of DOLFINX_MPC
 #
@@ -6,102 +6,14 @@
 from __future__ import annotations
 
 from mpi4py import MPI
-from petsc4py import PETSc
 
 import basix
 import dolfinx
-import dolfinx.fem.petsc
-import dolfinx.nls.petsc
 import numpy as np
 import pytest
 import ufl
-from dolfinx.la.petsc import create_vector
 
 import dolfinx_mpc
-
-
-class NonlinearMPCProblem(dolfinx.fem.petsc.NonlinearProblem):
-    def __init__(self, F, u, mpc, bcs=[], J=None, form_compiler_options={}, jit_options={}):
-        self.mpc = mpc
-        super().__init__(F, u, bcs=bcs, J=J, form_compiler_options=form_compiler_options, jit_options=jit_options)
-
-    def F(self, x: PETSc.Vec, F: PETSc.Vec):  # type: ignore
-        with F.localForm() as F_local:
-            F_local.set(0.0)
-        dolfinx_mpc.assemble_vector(self._L, self.mpc, b=F)
-
-        # Apply boundary condition
-        dolfinx_mpc.apply_lifting(
-            F,
-            [self._a],
-            bcs=[self.bcs],
-            constraint=self.mpc,
-            x0=[x],
-            scale=dolfinx.default_scalar_type(-1.0),
-        )
-        F.ghostUpdate(addv=PETSc.InsertMode.ADD, mode=PETSc.ScatterMode.REVERSE)  # type: ignore
-        dolfinx.fem.petsc.set_bc(F, self.bcs, x, -1.0)
-
-    def J(self, x: PETSc.Vec, A: PETSc.Mat):  # type: ignore
-        A.zeroEntries()
-        dolfinx_mpc.assemble_matrix(self._a, self.mpc, bcs=self.bcs, A=A)
-        A.assemble()
-
-
-class NewtonSolverMPC(dolfinx.cpp.nls.petsc.NewtonSolver):
-    def __init__(
-        self,
-        comm: MPI.Intracomm,
-        problem: NonlinearMPCProblem,
-        mpc: dolfinx_mpc.MultiPointConstraint,
-    ):
-        """A Newton solver for non-linear MPC problems."""
-        super().__init__(comm)
-        self.mpc = mpc
-        self.u_mpc = dolfinx.fem.Function(mpc.function_space)
-
-        # Create matrix and vector to be used for assembly of the non-linear
-        # MPC problem
-        self._A = dolfinx_mpc.cpp.mpc.create_matrix(problem.a._cpp_object, mpc._cpp_object)
-        self._b = create_vector(mpc.function_space.dofmap.index_map, mpc.function_space.dofmap.index_map_bs)
-
-        self.setF(problem.F, self._b)
-        self.setJ(problem.J, self._A)
-        self.set_form(problem.form)
-        self.set_update(self.update)
-
-    def update(self, solver: dolfinx.nls.petsc.NewtonSolver, dx: PETSc.Vec, x: PETSc.Vec):  # type: ignore
-        # We need to use a vector created on the MPC's space to update ghosts
-        self.u_mpc.x.petsc_vec.array = x.array_r
-        self.u_mpc.x.petsc_vec.axpy(-1.0, dx)
-        self.u_mpc.x.petsc_vec.ghostUpdate(
-            addv=PETSc.InsertMode.INSERT,  # type: ignore
-            mode=PETSc.ScatterMode.FORWARD,  # type: ignore
-        )  # type: ignore
-        self.mpc.homogenize(self.u_mpc)
-        self.mpc.backsubstitution(self.u_mpc)
-        x.array = self.u_mpc.x.petsc_vec.array_r
-        x.ghostUpdate(
-            addv=PETSc.InsertMode.INSERT,  # type: ignore
-            mode=PETSc.ScatterMode.FORWARD,  # type: ignore
-        )  # type: ignore
-
-    def solve(self, u: dolfinx.fem.Function):
-        """Solve non-linear problem into function u. Returns the number
-        of iterations and if the solver converged."""
-        n, converged = super().solve(u.x.petsc_vec)
-        u.x.scatter_forward()
-        return n, converged
-
-    @property
-    def A(self) -> PETSc.Mat:  # type: ignore
-        """Jacobian matrix"""
-        return self._A
-
-    @property
-    def b(self) -> PETSc.Vec:  # type: ignore
-        """Residual vector"""
-        return self._b
 
 
 @pytest.mark.skipif(
@@ -170,15 +82,27 @@ def test_nonlinear_poisson(poly_order):
 
         assert num_slaves_global > 0
         assert num_masters_global == num_slaves_global
-
-        problem = NonlinearMPCProblem(F, u, mpc, bcs=bcs, J=J)
-        solver = NewtonSolverMPC(mesh.comm, problem, mpc)
-        solver.atol = 1e1 * np.finfo(u.x.array.dtype).resolution
-        solver.rtol = 1e1 * np.finfo(u.x.array.dtype).resolution
+        tol = 1e1 * np.finfo(u.x.array.dtype).resolution
+        petsc_options = {
+            "snes_type": "newtonls",
+            "ksp_type": "preonly",
+            "pc_type": "lu",
+            "pc_factor_mat_solver_type": "mumps",
+            "snes_atol": tol,
+            "snes_rtol": tol,
+            "snes_linesearch_type": "none",
+            "ksp_error_if_not_converged": True,
+            "snes_error_if_not_converged": True,
+            "snes_monitor": None,
+        }
+        problem = dolfinx_mpc.NonlinearProblem(F, u, mpc=mpc, bcs=bcs, J=J, petsc_options=petsc_options)
+        # solver = NewtonSolverMPC(mesh.comm, problem, mpc)
 
         # Ensure the solver works with nonzero initial guess
         u.interpolate(lambda x: x[0] ** 2 * x[1] ** 2)
-        solver.solve(u)
+        _, converged, num_iterations = problem.solve()
+
+        print(f"Converged: {converged}, iterations: {num_iterations}")
         l2_error_local = dolfinx.fem.assemble_scalar(dolfinx.fem.form((u - u_soln) ** 2 * ufl.dx))
         l2_error_global = mesh.comm.allreduce(l2_error_local, op=MPI.SUM)